Autorization works

+++ ms-comment-rating/src/main/java/cl/govegan/ms_comment_rating/controller/CommentController.java
@@ -1,11 +1,14 @@
 package cl.govegan.ms_comment_rating.controller;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -18,24 +21,25 @@ import cl.govegan.ms_comment_rating.models.Comment;
 import cl.govegan.ms_comment_rating.services.CommentServices;
 
 
+
 @RestController
+@RequiredArgsConstructor
 @RequestMapping("/api/v1/comments")
 public class CommentController {
 
-
-   @Autowired
-    private CommentServices commentService;
+    private final CommentServices commentService;
 
    @GetMapping("/status")
    public ResponseEntity<String> status() {
       return ResponseEntity.ok("Comment service is up and running");
    }
+
    @PostMapping
     public ResponseEntity<String> createComment(@RequestBody Comment comment) {
       commentService.addComment(comment);
         return ResponseEntity.ok("Comment added correctly");
-
     }
+
     @GetMapping("/findByRecipeId")
     public ResponseEntity<Page<Comment>> findCommentsByRecipeId(
             @RequestParam String recipeId,
@@ -45,6 +49,7 @@ public class CommentController {
         Page<Comment> comments = commentService.findByRecipeId(recipeId, pageable);
         return ResponseEntity.ok(comments);
     }
+
     @GetMapping("/findByUsername")
     public ResponseEntity<Page<Comment>> findCommentsByUsername(
             @RequestParam String username,
@@ -69,6 +74,15 @@ public class CommentController {
     public ResponseEntity<String> deleteCommentByUsernameAndRecipeId(
             @RequestParam String recipeId,
             @RequestParam String username) {
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        String currentUsername = authentication.getName();
+
+        if (!currentUsername.equals(username)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("You can only delete your own comments.");
+        }
+
         boolean isDeleted = commentService.deleteCommentbyUsernameAndRecipeId(recipeId, username);
         if (isDeleted) {
             return ResponseEntity.ok("Comments deleted successfully.");
